Allow custom jenkins host when connecting to jenkins slave port via JNLP
The JNLP protocol provides the default jenkins hostname to the slave
which the slave uses by default to connect to the master server.
If the slave connects on another hostname/ip to the master it will
be redirected to the default jenkins hostname when the -connectTo
parameter for slave.jar is not set. The JENKINS_SLAVE_ADDRESS parameter
allows to set that value.